+++ app/app/(reception)/usage-logs/client-components/ScoreDisplayForm.tsx
