+++ app/app/(reception)/usage-logs/entities/HistoryListViewItemEntity.ts
