+++ app/app/(reception)/usage-logs/entities/HistoryPageEntity.ts
